pyproject.toml
[tool.poetry]
package-mode = false
name = "profile_epage"
version = "0.1.0"
description = "User CRUD "
authors = ["Denis <work80@mail.ru>"]
readme = "./README.md"

[tool.poetry.dependencies]
python = "^3.13.5"
python-dotenv = "^1.0.1"
django = "5.2.8"
djangorestframework = "^3.16.1"
adrf = "^0.1.12"
djangorestframework-simplejwt = "^5.5.1"
psycopg2-binary = "^2.9.11"
django-cors-headers = "^4.9.0"
postgres = "^4.0"
django-webpack-loader = "^3.2.2"
channels = {extras = ["daphne"], version = "^4.3.2"}
wagtail = "^7.2.1"
celery = "^5.5.3"
redis = "^7.1.0"
django-redis = "^6.0.0"
drf-yasg = "^1.21.11"
drf-spectacular = "^0.29.0"
bcrypt = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.5.0"
pytest = "^9.0.1"
pytest-cov = "^7.0.0"
pytest-django = "^4.11.1"
pytest-mock = "^3.15.1"
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
